Extract ManageShares Component and replace it for upselling Project List Share
