+++ app/components/shares/modal_body_component.rb
@@ -28,160 +28,21 @@
 
 module Shares
   class ModalBodyComponent < ApplicationComponent # rubocop:disable OpenProject/AddPreviewForViewComponent
-    include ApplicationHelper
-    include MemberHelper
     include OpTurbo::Streamable
     include OpPrimer::ComponentHelpers
 
     attr_reader :strategy,
-                :entity,
                 :errors
 
     def initialize(strategy:, errors: nil)
       super
 
       @strategy = strategy
-      @entity = strategy.entity
       @errors = errors
     end
 
     def self.wrapper_key
       "share_modal_body"
-    end
-
-    private
-
-    def project_scoped_entity?
-      entity.respond_to?(:project)
-    end
-
-    def insert_target_modified?
-      true
-    end
-
-    def insert_target_modifier_id
-      "op-share-dialog-active-shares"
-    end
-
-    def type_filter_options
-      if project_scoped_entity?
-        [
-          { label: I18n.t("sharing.filter.project_member"),
-            value: { principal_type: "User", project_member: true } },
-          { label: I18n.t("sharing.filter.not_project_member"),
-            value: { principal_type: "User", project_member: false } },
-          { label: I18n.t("sharing.filter.project_group"),
-            value: { principal_type: "Group", project_member: true } },
-          { label: I18n.t("sharing.filter.not_project_group"),
-            value: { principal_type: "Group", project_member: false } }
-        ]
-      else
-        [
-          { label: I18n.t("sharing.filter.user"), value: { principal_type: "User" } },
-          { label: I18n.t("sharing.filter.group"), value: { principal_type: "Group" } }
-        ]
-
-      end
-    end
-
-    def type_filter_option_active?(option) # rubocop:disable Metrics/AbcSize
-      principal_type_filter_value = current_filter_value(params[:filters], "principal_type")
-      project_member_filter_value = current_filter_value(params[:filters], "also_project_member")
-
-      return false if principal_type_filter_value.nil?
-      return false if project_scoped_entity? && project_member_filter_value.nil?
-
-      principal_type_checked =
-        option[:value][:principal_type] == principal_type_filter_value
-      membership_selected = !project_scoped_entity? ||
-        option[:value][:project_member] == ActiveRecord::Type::Boolean.new.cast(project_member_filter_value)
-
-      principal_type_checked && membership_selected
-    end
-
-    def role_filter_option_active?(option)
-      role_filter_value = current_filter_value(params[:filters], "role_id")
-
-      return false if role_filter_value.nil?
-
-      selected_role = strategy.available_roles.find { _1[:value] == option[:value] }
-
-      selected_role[:value] == role_filter_value.to_i
-    end
-
-    def filter_url(type_option: nil, role_option: nil)
-      return url_for([@entity, Member]) if type_option.nil? && role_option.nil?
-
-      args = {}
-      filter = []
-
-      filter += apply_role_filter(role_option)
-      filter += apply_type_filter(type_option)
-
-      args[:filters] = filter.to_json unless filter.empty?
-
-      url_for([@entity, Member, args])
-    end
-
-    def apply_role_filter(option)
-      current_role_filter_value = current_filter_value(params[:filters], "role_id")
-      filter = []
-
-      if option.nil? && current_role_filter_value.present?
-        # When there is already a role filter set and no new value passed, we want to keep that filter
-        filter = role_filter_for({ value: current_role_filter_value })
-      elsif option.present? && !role_filter_option_active?(option)
-        # Only when the passed filter option is not the currently selected one, we apply the filter
-        filter = role_filter_for(option)
-      end
-
-      filter
-    end
-
-    def role_filter_for(option)
-      [
-        { role_id: { operator: "=", values: [option[:value]] } }
-      ]
-    end
-
-    def apply_type_filter(option)
-      current_type_filter_value = current_filter_value(params[:filters], "principal_type")
-      current_member_filter_value = current_filter_value(params[:filters], "also_project_member")
-      filter = []
-
-      if option.nil? && current_type_filter_value.present? && current_member_filter_value.present?
-        # When there is already a type filter set and no new value passed, we want to keep that filter
-        value = { value: { principal_type: current_type_filter_value, project_member: current_member_filter_value } }
-        filter = type_filter_for(value)
-      elsif option.present? && !type_filter_option_active?(option)
-        # Only when the passed filter option is not the currently selected one, we apply the filter
-        filter = type_filter_for(option)
-      end
-
-      filter
-    end
-
-    def type_filter_for(option)
-      filter = [
-        { principal_type: { operator: "=", values: [option[:value][:principal_type]] } }
-      ]
-
-      if project_scoped_entity?
-        if ActiveRecord::Type::Boolean.new.cast(option[:value][:project_member])
-          filter.push({ also_project_member: { operator: "=", values: [OpenProject::Database::DB_VALUE_TRUE] } })
-        else
-          filter.push({ also_project_member: { operator: "=", values: [OpenProject::Database::DB_VALUE_FALSE] } })
-        end
-      end
-
-      filter
-    end
-
-    def current_filter_value(filters, filter_key)
-      return nil if filters.nil?
-
-      given_filters = JSON.parse(filters).find { |key| key.key?(filter_key) }
-      given_filters ? given_filters[filter_key]["values"].first : nil
     end
   end
 end
