+++ app/models/sharing_strategies/project_query_strategy.rb
@@ -105,6 +105,14 @@ module SharingStrategies
       end
     end
 
+    def manage_shares_component(modal_content:, errors:)
+      if EnterpriseToken.allows_to?(:project_list_sharing)
+        super
+      else
+        Shares::ProjectQueries::UpsaleComponent.new(modal_content:)
+      end
+    end
+
     private
 
     def virtual_owner_share
